enable awesome build.sh

+++ build.sh
@@ -1,19 +1,41 @@
 #!/bin/sh -e
-
 HOME_DIR=`pwd`
 
+while getopts "s:o:r:d:y" optname
+do
+    case "$optname" in
+    "s")
+	SYNC_PORT=$OPTARG
+        ;;
+    "o")
+	OPENFLOW_PORT=$OPTARG
+        ;;
+    "r")
+	REST_PORT=$OPTARG
+        ;;
+    "d")
+	ONEX_DAEMON_PORT=$OPTARG
+        ;;
+    "y")
+	yes_mode="yes"	
+	;;
+    *)
+        echo "Unknown error while processing options"
+        ;;
+    esac
+done
+
 # for floodlight only
-SYNC_PORT=6788
-OPENFLOW_PORT=6633
-REST_PORT=8080
-DP_PATH='\/tmp\/123'
+
+SYNC_PORT=${SYNC_PORT-6700}
+OPENFLOW_PORT=${OPENFLOW_PORT-6633}
+REST_PORT=${REST_PORT-8080}
+ONEX_DAEMON_PORT=${ONEX_DAEMON_PORT-7888}
+
+temp=`mktemp -d /tmp/XXXXXX | cut -d '/' -f 3`
 CONTROLLER_IP=1
-
+DP_PATH='\/tmp\/'$temp
 # COMMON
-ONEX_DAEMON_PORT=7888
-
-# for ONEx
-GATEWAY_SERVER_PORT=7999
 
 # dir and files
 FLOODLIGHT_ROOT_DIR='/Users/Fan/dev/floodlight'
@@ -21,18 +43,7 @@ ONEX_ROOT_DIR="/Users/Fan/dev/ONEx/ONEx"
 # files
 FLOODLIGHT_PROP_FILE=$FLOODLIGHT_ROOT_DIR"/src/main/resources/floodlightdefault.properties"
 FLOODLIGHT_TEST_SETTING_FILE=$FLOODLIGHT_ROOT_DIR"/src/main/java/ONExProtocol/TestSetting.java"
-ONEX_CONFIG_FILE=$ONEX_ROOT_DIR"/src/config.properties"
 
-echo "ONEx config.properties>>"
-echo "GATEWAY_SERVER_PORT=$GATEWAY_SERVER_PORT"
-l=`grep -n GATEWAY_SERVER_PORT $ONEX_CONFIG_FILE | awk -F : '{print $1}'`
-sed -i.bak $l's/.*/GATEWAY_SERVER_PORT '$GATEWAY_SERVER_PORT'/' $ONEX_CONFIG_FILE
-
-echo "DAEMON_PORT=$ONEX_DAEMON_PORT"
-l=`grep -n DAEMON_PORT $ONEX_CONFIG_FILE | awk -F: '{print $1}'`
-sed  -i.bak $l's/.*/DAEMON_PORT '$ONEX_DAEMON_PORT'/' $ONEX_CONFIG_FILE
-
-echo "floodlightdefault.properties>>"
 echo "+openflowport=$OPENFLOW_PORT"
 javal=`grep -n openflowport $FLOODLIGHT_PROP_FILE`
 javaln=`echo $javal | awk -F : '{ print $1;}'`
@@ -52,7 +63,6 @@ javalnew=`echo $javal | cut -d: -f2- | awk -F = '{print $1"="}'`$DP_PATH
 sed -i.bak $javaln's/.*/'"$javalnew"'/' $FLOODLIGHT_PROP_FILE
 echo ""
 
-echo "TestSetting.java>>"
 echo "+controllerIP=$CONTROLLER_IP"
 javal=`grep -n controllerIP $FLOODLIGHT_TEST_SETTING_FILE`
 javaln=`echo $javal | awk -F : '{print $1;}'`
@@ -84,25 +94,47 @@ javalnew=`echo $javal | cut -d: -f2- | awk -F = '{print $1"="}'`"$ONEX_DAEMON_PO
 sed -i.bak $javaln's/.*/'"$javalnew"'/' $FLOODLIGHT_TEST_SETTING_FILE
 
 confirm(){
-	clear && vim $ONEX_CONFIG_FILE
-	clear && vim $FLOODLIGHT_PROP_FILE
-	clear && vim $FLOODLIGHT_TEST_SETTING_FILE
+	vim $FLOODLIGHT_PROP_FILE
+	vim $FLOODLIGHT_TEST_SETTING_FILE
 }
 
 build(){
-	clear && echo "building..."
-	cd $FLOODLIGHT_ROOT_DIR
-	ant dist && cp $FLOODLIGHT_ROOT_DIR/target/floodlight.jar \
-	$HOME_DIR/"floodlight-$ONEX_DAEMON_PORT(daemon)-$OPENFLOW_PORT(openflow).jar"
+	echo "building..."
+	TARGET_DIR=$HOME_DIR/release
+	mkdir -p $TARGET_DIR
 	
-	cd $ONEX_ROOT_DIR
-	ant && cp $ONEX_ROOT_DIR/build/jar/*.jar \
-		$HOME_DIR/"daemon-$ONEX_DAEMON_PORT(daemon)-$OPENFLOW_PORT(openflow).jar"
-	cp $ONEX_ROOT_DIR/lib/*.so $HOME_DIR
-	cp $ONEX_ROOT_DIR/lib/*.dll $HOME_DIR
-	cp $ONEX_ROOT_DIR/lib/*.dylib $HOME_DIR
-
+	cd $FLOODLIGHT_ROOT_DIR
+	rm -rf target
+	ant && cp $FLOODLIGHT_ROOT_DIR/target/floodlight.jar \
+	$TARGET_DIR/"floodlight-o$OPENFLOW_PORT-d$ONEX_DAEMON_PORT.jar"
+	
+	if test $yes_mode
+	then yn="Yes"
+	else
+		read -p "recompile ONExService?[y/n]" yn
+	fi
+	case $yn in 
+	[Yy]* )		
+		cd $ONEX_ROOT_DIR
+		ant && cp $ONEX_ROOT_DIR/build/jar/*.jar \
+			$TARGET_DIR/"ONExService-pd-ps.jar"
+	
+		cd $ONEX_ROOT_DIR/lib/;
+		for i in `find . -name "libsigar*"`;
+		do
+			cp $i $TARGET_DIR
+		done;
+		break;;
+	*) echo "Bye!";
+	esac
 }
+
+if test $yes_mode
+then 
+	build;
+	echo "Enjoy!"
+	exit;
+fi
 
 while true; do
     read -p "Confirm? [y/n]" yn
@@ -122,5 +154,4 @@ while true; do
     esac
 done
 
-
-
+echo "Enjoy!"
